export command executions to CommandExecutors
